+++ src/screens/Donate/content.js
@@ -1,0 +1,32 @@
+export const donate1Head = 'Unterstütze DEMOCRACY';
+export const donate1Text = `Unsere Arbeit ist unabhängig, überparteilich, allgemeinnützig und nicht-kommerziell –
+von Menschen für Menschen.
+
+Wenn Du Interesse daran hast, dass das Projekt nachhaltig Bestand hat, bitten wir Dich an der unabhängigen Crowd-Finanzierung von DEMOCRACY per Dauerauftrag mitzuwirken. Pate/in können Sie schon ab 1€ monatlich werden.
+
+Spendenstand vom 16.09.2018`;
+export const donate2Head = 'Spendenkonto für Daueraufträge oder einmalige Spenden';
+export const donateList1Head = 'Kontoinhaber:';
+export const donateList1Text = 'DEMOCRACY Deutschland e.V.';
+export const donateList2Head = 'IBAN:';
+export const donateList2Text = 'DE33 5003 1000 1049 7560 00';
+export const donateList3Head = 'BIC:';
+export const donateList3Text = 'TRODDEF1';
+export const donate3Text1 = `Für den Verein DEMOCRACY Deutschland e.V. ist mit Bescheid des Finanzamt Göttingen vom 23.08.2017 die Einhaltung der satzungsmäigen Voraussetzungen nach den §§ 51, 59, 60 und 61 AO festgestellt worden (`;
+export const donate3Text2 = 'Nachweis der Gemeinnützigkeit';
+export const donate3Link1 =
+  'https://www.democracy-deutschland.de/files/download/Gemeinnutzigkeit_DEMOCRACY.pdf';
+export const donate3Text3 = '; ';
+export const donate3Text4 = 'Satzung';
+export const donate3Link2 =
+  'https://www.democracy-deutschland.de/files/download/Satzung_DEMOCRACY.pdf';
+export const donate3Text5 = `). Deine Spenden sind daher steuerlich abzugsfähig. Für Spenden unter 200€ jährlich genügt der Zahlungsnachweis (z.B. Kontoauszug) zusammen mit unserem `;
+export const donate3Text6 = 'Vereinfachten Zuwendungsnachweis';
+export const donate3Link3 =
+  'https://www.democracy-deutschland.de/files/download/Vereinfachter_Zuwendungsnachweis.pdf';
+export const donate3Text7 = `. Wenn Du eine separate Zuwendungsbestätigung benötigst, schreib uns bitte eine E-Mail oder gib in Deiner Überweisung Deine Wohnadresse an.`;
+export const donate4Head = 'Spenden via PayPal';
+export const donate4Text = `1,5% Deines Spendenbetrags + 0,35 EUR pro Transaktion verbleiben bei PayPal.
+Bitte bedenke, dass von einer monatlichen 1€-Spende insofern nur 0,63€ bei uns ankommen. 
+
+Damit Deine Spende zu 100% wirksam wird und für Dich jederzeit kontrollierbar ist, empfehlen wir einen Dauerauftrag á Konto einzurichten.`;