+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.10.1] - 2018-09-XX
+## [0.10.2] - 2018-09-19
+### Added
+- [SMS Verification] Success Screen
+- [Donate] Donate informations
+- [FAQ] New FAQ page
+- [BurgerMenu] New Texts for all Menus
+
+## [0.10.1] - 2018-09-17
 ### Added
 - Beta End Nag Screen